Move small bits of code out of Hyperion
