+++ src/mx_bluesky/hyperion/experiment_plans/change_aperture_then_move_plan.py
@@ -6,8 +6,8 @@ from dodal.devices.smargon import Smargon, StubPosition
 
 from mx_bluesky.common.utils.log import LOGGER
 from mx_bluesky.common.utils.tracing import TRACER
+from mx_bluesky.common.xrc_result import XRayCentreResult
 from mx_bluesky.hyperion.device_setup_plans.manipulate_sample import move_x_y_z
-from mx_bluesky.hyperion.experiment_plans.common.xrc_result import XRayCentreResult
 from mx_bluesky.hyperion.parameters.gridscan import HyperionSpecifiedThreeDGridScan
